Fixing up the about me page
Tried a few different design patterns and methods to create the about me section, might have to look at different pages for inspiration as to how it should look while containing the things I am looking for. Then I also added a dockermake file to start trying to containerize the application while also adding the backend folder system to start being able to host this thing through an actual server and proxy the two together. Though that is for another day in time...